refactor: remove `assetAmount` state
